refactor: use milliseconds for block timestamp (#156)
* support uint48 serialization

* block serde

* regenerate genesis block

* fix consensus scheduler

* fix tests

* style: resolve style guide violations

---------

+++ packages/crypto-block/source/deserializer.ts
@@ -103,10 +103,10 @@ export class Deserializer implements Contracts.Crypto.IBlockDeserializer {
 			length: this.blockSerializer.headerSize(),
 			schema: {
 				version: {
-					type: "uint32",
+					type: "uint8",
 				},
 				timestamp: {
-					type: "uint32",
+					type: "uint48",
 				},
 				height: {
 					type: "uint32",
